[Unticketed] Fetching Project Query Optimization (#1618)
* removed shipping rules from reward fragment, as it is called later in the pipeline when the user selects an addon or shipping rule.

no tests. Needs a smoke test.

* added tests

* moved the rewards from the fetch project template to the fetch rewards template

* forgot to add back in no rewards after it was moved to the view model
