+++ KsApi/models/graphql/adapters/Project+FetchProjectQueryData.swift
@@ -45,19 +45,6 @@ extension Project {
         Reward.reward(from: fragment)
       }
 
-    let rewards = data.project?.rewards?.nodes?
-      .compactMap { node -> GraphAPI.RewardFragment? in
-        guard let rewardFragment = node?.fragments.rewardFragment else { return nil }
-
-        return rewardFragment
-      }
-      .compactMap { fragment in
-        Reward.reward(from: fragment)
-      } ?? []
-
-    let emptyRewards = [noRewardReward(from: data.project?.fragments.projectFragment)]
-    let allRewards = emptyRewards + rewards
-
     var projectBackingId: Int?
 
     if let backingId = data.project?.backing?.id {
@@ -68,7 +55,7 @@ extension Project {
       let fragment = data.project?.fragments.projectFragment,
       let project = Project.project(
         from: fragment,
-        rewards: allRewards,
+        rewards: [noRewardReward(from: fragment)],
         addOns: addOns,
         backing: nil,
         currentUserChosenCurrency: data.me?.chosenCurrency
@@ -89,19 +76,6 @@ extension Project {
         Reward.reward(from: fragment)
       }
 
-    let rewards = data.project?.rewards?.nodes?
-      .compactMap { node -> GraphAPI.RewardFragment? in
-        guard let rewardFragment = node?.fragments.rewardFragment else { return nil }
-
-        return rewardFragment
-      }
-      .compactMap { fragment in
-        Reward.reward(from: fragment)
-      } ?? []
-
-    let emptyRewards = [noRewardReward(from: data.project?.fragments.projectFragment)]
-    let allRewards = emptyRewards + rewards
-
     var projectBackingId: Int?
 
     if let backingId = data.project?.backing?.id {
@@ -112,7 +86,7 @@ extension Project {
       let fragment = data.project?.fragments.projectFragment,
       let project = Project.project(
         from: fragment,
-        rewards: allRewards,
+        rewards: [noRewardReward(from: fragment)],
         addOns: addOns,
         backing: nil,
         currentUserChosenCurrency: data.me?.chosenCurrency
