+++ KsApi/queries/FetchBackingQuery.graphql
@@ -1,4 +1,4 @@
-query FetchBacking($id: ID!, $withStoredCards: Boolean!) {
+query FetchBacking($id: ID!, $withStoredCards: Boolean!, $includeShippingRules: Boolean!) {
   backing(id: $id) {
     addOns {
       nodes {
