+++ KsApi/queries/FetchProjectBySlugQuery.graphql
@@ -1,4 +1,4 @@
-query FetchProjectBySlug($slug: String!, $withStoredCards: Boolean!) {
+query FetchProjectBySlug($slug: String!, $withStoredCards: Boolean!, $includeShippingRules: Boolean!) {
   me {
     chosenCurrency
   }
@@ -11,11 +11,6 @@ query FetchProjectBySlug($slug: String!, $withStoredCards: Boolean!) {
     }
     backing {
       id
-    }
-    rewards {
-      nodes {
-        ...RewardFragment
-      }
     }
   }
 }
