+++ KsApi/queries/FetchUserBackings.graphql
@@ -1,4 +1,4 @@
-query FetchUserBackings($status: BackingState!, $withStoredCards: Boolean!) {
+query FetchUserBackings($status: BackingState!, $withStoredCards: Boolean!, $includeShippingRules: Boolean!) {
   me {
     backings(status: $status) {
       nodes {
